requirements.txt
dbt==1.0.0.37.15
dbt-core==1.8.6
dbt-trino==1.8.1
dbt-postgres==1.8.2
dbt-vertica==1.0.4
dbt-adapters==1.4.1
astronomer-cosmos==1.6.0
apache-airflow-providers-trino==5.8.0
apache-airflow-providers-postgres==5.13.0
psycopg2-binary==2.9.7
python-dotenv==1.0.1